+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/GeneticsResequenced.kt
@@ -6,6 +6,8 @@ import dev.aaronhowser.mods.geneticsresequenced.registry.ModRegistries
 import net.neoforged.fml.ModContainer
 import net.neoforged.fml.common.Mod
 import net.neoforged.fml.config.ModConfig
+import net.neoforged.neoforge.client.gui.ConfigurationScreen
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import thedarkcolour.kotlinforforge.neoforge.forge.MOD_BUS
@@ -22,6 +24,9 @@ class GeneticsResequenced(
 
     init {
         ModRegistries.register(MOD_BUS)
+
+        val screenFactory = IConfigScreenFactory { container, screen -> ConfigurationScreen(container, screen) }
+        modContainer.registerExtensionPoint(IConfigScreenFactory::class.java, screenFactory)
 
         modContainer.registerConfig(ModConfig.Type.CLIENT, ClientConfig.CONFIG_SPEC)
         modContainer.registerConfig(ModConfig.Type.SERVER, ServerConfig.CONFIG_SPEC)